% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqua.R
\docType{class}
\name{Aqua}
\alias{Aqua}
\alias{aqua}
\title{Get a Aqua object}
\format{\code{\link{R6Class}} object.}
\usage{
Aqua$new(session = ssion, use.defaults = T)
}
\value{
Object of R6 class which hold connection information
}
\description{
\code{Aqua} generates a S6 class
}
\details{
Aqua provides methods to cache the result of functions calls.
}
\section{Methods}{

These methods are accessible:

\describe{
\item{\code{store(file = NULL)}}{Saves cached values to file}
}
}

\examples{
aqua = Aqua$new(session = ssion, use.defaults = T)
aqua$queries = NULL #reset all queries
aqua$filters = NULL #reset all filters
Aqua$listAggregation(ssion)
Aqua$annotations(session = ssion, variation_ids = 1, organism_id = 1, model = "Vep", attributes = NULL)
akeys = c(
 'aggregation_variation_call:variation_calls.variation_id',
	'aggregation_variation_call:variation_calls.dp',
	'aggregation_annovar:annovars.organism_id',
	'aggregation_annovar:annovars.ensembl_annotation',
	'aggregation_annovar:annovars.ensembl_gene'
)
# fkey
qkeys = list(
	"query_consequence:consequence" = c("missense_variant", "stop_gained"),
	#"query_consequence:consequence[]" = "missense_variant",
	#"query_consequence:consequence[]" = "stop_gained",
	"query_variation_call:read_depth" = 100,
	"query_simple_population:population_frequency" = 0.01
)
fkeys = c('vep_filter_consequence:consequence_severe', 'filter_variation_call:vcdp', 'vep_filter_variant:vep_onekg', 'vep_filter_variant:vep_exac')

eg = EntityGroup$new(3443, ssion)
sc = eg$samples()
aqua = Aqua$new(ssion)
aqua$query_and_aggregate(sc)
res = aqua$query(sc)
agg = aqua$aggregate(res)
}
\seealso{
\code{\link{SnupySession}}, \code{\link{snupy.query}}, \code{\link{snupy.request}}

Other cache: \code{\link{SnupyDataCollection}},
  \code{\link{SnupyData}}, \code{\link{SnupyMemory}}
}
\author{
Sebastian Ginzel <sginze2s@inf.h-brs.de>
}
\keyword{datasets}
