% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{buildStatement}
\alias{buildStatement}
\title{Buid A SQL statement}
\usage{
buildStatement(table, columns = "*", joins = list(), limit = "",
  group.by = "", having = "", order.by = "", ...)
}
\description{
Buid A SQL statement
}
