% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snupy_data_collection_classes.R
\docType{data}
\name{EntityCollection}
\alias{EntityCollection}
\title{EntitysCollection}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EntityCollection
}
\description{
EntitysCollection
}
\keyword{datasets}
