% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snupy_data_collection_classes.R
\docType{data}
\name{EntityGroupCollection}
\alias{EntityGroupCollection}
\title{EntityGroupsCollection}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EntityGroupCollection
}
\description{
EntityGroupsCollection
}
\keyword{datasets}
