% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_group.R
\docType{data}
\name{EntityGroupTag}
\alias{EntityGroupTag}
\title{SnupyEntityGroupTag}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EntityGroupTag
}
\description{
SnupyEntityGroupTag
}
\keyword{datasets}
