% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snupy_data_collection_classes.R
\docType{data}
\name{ExperimentCollection}
\alias{ExperimentCollection}
\title{Experiment}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ExperimentCollection
}
\description{
Experiment
}
\keyword{datasets}
