% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snupy_graph_helper.R
\name{getIdentity}
\alias{getIdentity}
\alias{identity}
\alias{i}
\title{Generate identity matrix}
\usage{
getIdentify(d, diag.value = 1)
}
\arguments{
\item{d}{Dimension of identity matrix.}

\item{diag.value}{Value on the diagonal of the matrix. Default: 1}
}
\value{
Returns a sparseMatrix.
}
\description{
\code{getIdentity} createsn identity matrix with the given dimension as a sparseMatrix.
}
\details{
This method creates a square matrix for the dimension given by d.\cr
}
\seealso{
\code{\link{graph_diffusion}}

Other matrix operations: \code{\link{.graph.diffusion.converges}},
  \code{\link{maxEigen}}
}
\author{
Sebastian Ginzel <sginze2s@inf.h-brs.de>
}
