% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snupy_graph_helper.R
\name{maxEigen}
\alias{maxEigen}
\alias{maximum}
\alias{eigenvalue}
\alias{maxeigen}
\alias{max.eigen}
\title{Power method to approximate the maximum eigenvalue of a matrix}
\usage{
maxEigen(m, e = 10^-16, max.iter = 1000)
}
\arguments{
\item{m}{a numeric matrix.}

\item{e}{error threhold for the angle between the vectors before and after the matrix multiplication. Default: 10^-16}

\item{max_iter}{maximum number of iterations to perform. Default: 1000}
}
\value{
max.eigen return the approximated maximum eigenvalue of the matrix or NA in case convergence was not achieved after max.iter iterations.
}
\description{
\code{maxEigen} performs maximum eigenvalue determination.
}
\details{
The method starts with a vector of 1's and multiplies it continously with the given matrix. When the angle between the input and the result vector converges we determine the eigenvalue by the length ratio between the two vectors. \cr
}
\seealso{
\code{\link{graph_diffusion}}

Other matrix operations: \code{\link{.graph.diffusion.converges}},
  \code{\link{getIdentity}}
}
\author{
Sebastian Ginzel <sginze2s@inf.h-brs.de>
}
