% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\docType{class}
\name{Sample}
\alias{Sample}
\alias{sample}
\title{Get and Sample object}
\format{\code{\link{R6Class}} object.}
\usage{
Sample$new(experiment_id=2)
Sample$new(experiment=Experiment$new(2))
Sample$new(id=123)
}
\arguments{
\item{filepath}{a path to a .r file that contains the config}

\item{connect}{logical value whether ssh tunnel should be connected or not.}
}
\value{
Object of R6 class to use Samples
}
\description{
Get and Sample object
}
\details{
The file containig the configuration should set the following variables\cr
\describe{
\item{name}{logical value if debug information should be shown}
\item{tags}{tags linked to this entity group}
}
}
\section{Fields}{

\describe{
\item{\code{name}}{Name of the \code{\link{Sample}}
#' @section Methods:
 \describe{
  \item{\code{has_tag(Tag$new(id=1))}}{Returns true if the Sample has is associated with Tag.}
}}
}}

\section{Methods}{

}

\examples{
eg <- Sample$new(experiment_id=2)
}
\seealso{
\code{\link{Experiment}}, \code{\link{Entity}}, \code{\link{Specimen}}, \code{\link{Sample}} and \code{\link{VcfFile}}

Other data: \code{\link{EntityGroup}},
  \code{\link{Entity}}, \code{\link{Experiment}},
  \code{\link{GenePanel}}, \code{\link{SnupyGraph}},
  \code{\link{SnupyTag}}, \code{\link{SpecimenProbe}},
  \code{\link{VariantDataSet}}, \code{\link{VcfFile}}
}
\author{
Sebastian Ginzel <sginze2s@inf.h-brs.de>
}
\keyword{datasets}
