% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{snupy.query}
\alias{snupy.query}
\alias{query}
\title{Sends a SQL query to the SNUPY API}
\usage{
snupy.query(statement, session, primary.key = NULL, method = c("curl",
  "wget"), format = c("data.table", "tibble"), extract = NULL,
  extract.name = NULL, cache.use = FALSE, cache.overwrite = FALSE,
  http.format = "csv")
}
\arguments{
\item{statement}{SQL query to execute.}

\item{session}{session variable as returned by \code{\link{snupy.createSession}}}

\item{primary.key}{can be character vector, or method. If it is a character vector
the key of the data.table object will be set to primary.key. If a method is passed
it should accept each record of the response as a parameter and return a key to be used.
When 'md5' is used as a parameter the method will calculate the md5 sum of each record
and use it as a key. Requires openssl package.}

\item{method}{use either wget command or curllib}

\item{format}{by default a data.table object is returned, but tibbles can be used as well.}
}
\value{
Will return the result of the query as a data.table (default) or tibble.
}
\description{
\code{snupy.query} sends a SQL query.
}
\details{
currently using curl also enables you to use a proxy with authentification.
}
\seealso{
\code{\link{snupy.createSession}}
}
\author{
Sebastian Ginzel <sginze2s@inf.h-brs.de>
}
