% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snupy_graph_helper.R
\name{snupy_graph.getA}
\alias{snupy_graph.getA}
\title{Helper methods for snupy_graph diffusions}
\usage{
snupy_graph.getA(grph, A.attr = "weight")
}
\arguments{
\item{filepath}{a path to a .r file that contains the config}

\item{connect}{logical value whether ssh tunnel should be connected or not.}
}
\value{
Object of R6 class to use Entitys
}
\description{
Helper methods for snupy_graph diffusions
}
\details{
The file containig the configuration should set the following variables\cr
}
\section{Fields}{

\describe{
\item{\code{name}}{Name of the \code{\link{Entity}}
#' @section Methods:
 \describe{
  \item{\code{has_tag(Tag$new(id=1))}}{Returns true if the Entity has is associated with Tag.}
}}
}}

\seealso{
\code{\link{SnupyGraph}}

Other graph: \code{\link{snupy_graph.diffusion.matrix}}
}
\author{
Sebastian Ginzel <sginze2s@inf.h-brs.de>
}
