% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snupy_data.R
\docType{class}
\name{SnupyData}
\alias{SnupyData}
\alias{session}
\title{Get a SnupyData object}
\format{\code{\link{R6Class}} object.}
\usage{
SnupyData$new(cache.path = NULL, verbose = FALSE)
}
\arguments{
\item{cache.path}{path to file that is used to store the cache if requested.}

\item{verbose}{logical value}
}
\value{
Object of R6 class which hold connection information
}
\description{
\code{SnupyData} generates a S6 class
}
\details{
SnupyData provides methods to cache the result of functions calls.
}
\section{Methods}{

These methods are accessible:

\describe{
\item{\code{store(file = NULL)}}{Saves cached values to file}
\item{\code{load(file = NULL)}}{loads cached values from file }
\item{\code{reset()}}{resets cache and performs gc(reset=T)}
\item{\code{show()}}{Shows list of cache elements}
\item{\code{size()}}{Shows list of cache elements}
\item{\code{call()}}{Shows current cache size.}
\item{\code{add(key, value)}}{add key/value pair to cache}
\item{\code{get_or_add(key, value)}}{sets key/value pair or returs value of key, if already set.}
\item{\code{get(key)}}{returns key}
\item{\code{get_key(FUN, ..., memoize.name = NULL, memoize.ignore.parameter = c())}}{
Returns a key for a call to FUN(...), if memoize.name is set, it return that value.
memoize.ignore.parameter can be used to excluse parameters from ...
(such as time stamps or session ids), which change during calls, but don't affect the result.}
\item{\code{cache(FUN, ..., memoize.force = FALSE, memoize.name = NULL, memoize.ignore.parameter = c())}}{
Adds an entry to the cache. CACHE[get_key(FUN, ..., memoize.name = NULL, memoize.ignore.parameter = c())] = FUN(...)
if it was not computed before. If it was the cache result is returned.
\cr
If memoize.force = TRUE the computation is forced.
}
}
}

\examples{
> library(snupy)
> mem = SnupyData$new(verbose = T)
> system.time(mem$cache(Sys.sleep, 2))
Computing new value for FUN.db8e490a925a60e62212cefc7674ca02
Returning value for FUN.db8e490a925a60e62212cefc7674ca02
User      System elapsed
0.00        0.00        2.03
> system.time(mem$cache(Sys.sleep, 2))
Returning value for FUN.db8e490a925a60e62212cefc7674ca02
User      System elapsed
0.03        0.00        0.03
}
\seealso{
\code{\link{SnupySession}}, \code{\link{snupy.query}}, \code{\link{snupy.request}}

Other cache: \code{\link{Aqua}},
  \code{\link{SnupyDataCollection}},
  \code{\link{SnupyMemory}}
}
\author{
Sebastian Ginzel <sginze2s@inf.h-brs.de>
}
\keyword{datasets}
