% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snupy_graph.R
\docType{class}
\name{SnupyGraph}
\alias{SnupyGraph}
\alias{graph}
\title{SnupyGraph Object}
\format{\code{\link{R6Class}} object.}
\usage{
grph = SnupyGraph$new("human", 750, "10")
}
\arguments{
\item{organism}{Can be human or mouse, or any valid StringDB taxon id.}

\item{threshold}{Threshold for StringDB links (default: 750)}

\item{version}{StringDB version (default: "10")}

\item{session}{a SnupySession object.}
}
\value{
Object of R6 class to use SnupyGrtaph
}
\description{
\code{SnupyGraph} generates a S6 class
}
\details{
The class wraps a StringDB object and provides additional methods to perform diffusion and mapping between variants (\code{\link{SnupyData}}) and the PPI network.
}
\section{Fields}{

\describe{
\item{\code{version}}{Version of the StringDB}
}}

\section{Methods}{

 \describe{
  \item{\code{diffusion()}}{Returns a diffusion matrix for the graph.}
  \item{\code{plot.diffusion()}}{Returns a diffusion based plot.}
  \item{\code{rank.diffusion()}}{Returns a diffusion based rank.}
}
}

\examples{
grph = SnupyGraph$new("human", 750, "10")
}
\seealso{
\code{\link{Experiment}}, \code{\link{EntityGroup}}, \code{\link{Entity}}, \code{\link{Specimen}}, \code{\link{Sample}} and \code{\link{VcfFile}}

Other data: \code{\link{EntityGroup}},
  \code{\link{Entity}}, \code{\link{Experiment}},
  \code{\link{GenePanel}}, \code{\link{Sample}},
  \code{\link{SnupyTag}}, \code{\link{SpecimenProbe}},
  \code{\link{VariantDataSet}}, \code{\link{VcfFile}}
}
\author{
Sebastian Ginzel <sginze2s@inf.h-brs.de>
}
\keyword{datasets}
