% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snupy_tag.R
\docType{class}
\name{SnupyTag}
\alias{SnupyTag}
\alias{entity_group}
\title{Get a SnupyTag object}
\format{\code{\link{R6Class}} object.}
\usage{
EntityGroup$new(experiment_id=2)
EntityGroup$new(experiment=Experiment$new(2))
EntityGroup$new(id=123)
}
\value{
Object of R6 class to use EntityGroups
}
\description{
\code{SnupyTag} generates a S6 class
}
\details{
The file containig the configuration should set the following variables\cr
\describe{
\item{name}{logical value if debug information should be shown}
\item{tags}{tags linked to this entity group}
}
}
\section{Fields}{

\describe{
\item{\code{name}}{Name of the \code{\link{EntityGroup}}
#' @section Methods:
 \describe{
  \item{\code{has_tag(Tag$new(id=1))}}{Returns true if the EntityGroup has is associated with Tag.}
}}
}}

\examples{
eg <- EntityGroup$new(experiment_id=2)
}
\seealso{
\code{\link{Experiment}}, \code{\link{Entity}}, \code{\link{Specimen}}, \code{\link{Sample}} and \code{\link{VcfFile}}

Other data: \code{\link{EntityGroup}},
  \code{\link{Entity}}, \code{\link{Experiment}},
  \code{\link{GenePanel}}, \code{\link{Sample}},
  \code{\link{SnupyGraph}}, \code{\link{SpecimenProbe}},
  \code{\link{VariantDataSet}}, \code{\link{VcfFile}}
}
\author{
Sebastian Ginzel <sginze2s@inf.h-brs.de>
}
\keyword{datasets}
