% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snupy_tag.R
\docType{data}
\name{SnupyTagCollection}
\alias{SnupyTagCollection}
\title{SnupyTags}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SnupyTagCollection
}
\description{
SnupyTags
}
\keyword{datasets}
