% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specimen.R
\docType{data}
\name{SpecimenProbeTag}
\alias{SpecimenProbeTag}
\title{SnupySpecimenTag}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SpecimenProbeTag
}
\description{
SnupySpecimenTag
}
\keyword{datasets}
