% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_data_set.R
\docType{class}
\name{VariantDataSet}
\alias{VariantDataSet}
\alias{VariantCallSet}
\title{Use a VariantDataSet module to filter variation calls}
\format{\code{\link{R6Class}} object.}
\usage{
VariantDataSet$new(sample_collection_or_sample_ids, session)
}
\arguments{
\item{collection_or_sample_ids}{a collection of ids or a vector of sample ids}

\item{session}{session object}
}
\value{
VariantDataSet
}
\description{
\code{VariantDataSet} generates a S6 class
}
\details{
Variant calls are retrieved seperately for each sample, when needed.
}
\section{Fields}{

\describe{
\item{\code{session()}}{session object}

\item{\code{variation_calls()}}{table of variation calls}

\item{\code{samples()}}{returns the sample objects}

\item{\code{variants}}{list of variant ids}

\item{\code{samples_ids}}{returns a vector of sample_ids}
}}

\section{Methods}{

 \describe{
  \item{\code{regions(include.alterations = F, as.ranges = T)}}{Returns the coordinates of the variants in the dataset. Optionally as GRanges objects(not implemented)}
  \item{\code{variation.matrix(attribute = c("present", "baf"), min.dp = 100, col.row.name = "id")}}{A sample_id-variation matrix, possibly filtered for read depth. Cells are binary (present) or are set to the variation frequency.}
  \item{\code{aqua(aqua=NULL)}}{performs aqua aggregation for variants in the set, in case the variants are already filtered.}
  \item{\code{+,-,&,|}}{Performs set operations on variant datasets. Returns a new variant dataset each time.}
}
}

\examples{
VariantDataSet$new(c(1,2), ssion)
}
\seealso{
\code{\link{Experiment}}, \code{\link{Entity}}, \code{\link{Specimen}}, \code{\link{Sample}} and \code{\link{VcfFile}}

Other data: \code{\link{EntityGroup}},
  \code{\link{Entity}}, \code{\link{Experiment}},
  \code{\link{GenePanel}}, \code{\link{Sample}},
  \code{\link{SnupyGraph}}, \code{\link{SnupyTag}},
  \code{\link{SpecimenProbe}}, \code{\link{VcfFile}}
}
\author{
Sebastian Ginzel <sginze2s@inf.h-brs.de>
}
\keyword{datasets}
