% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_file.R
\docType{class}
\name{VcfFile}
\alias{VcfFile}
\title{Get and VcfFile object}
\format{\code{\link{R6Class}} object.}
\usage{
VcfFile$new(experiment_id=2)
VcfFile$new(experiment=Experiment$new(2))
VcfFile$new(id=123)
}
\arguments{
\item{filepath}{a path to a .r file that contains the config}

\item{connect}{logical value whether ssh tunnel should be connected or not.}
}
\value{
Object of R6 class to use VcfFiles
}
\description{
Get and VcfFile object
}
\details{
The file containig the configuration should set the following variables\cr
\describe{
\item{name}{logical value if debug information should be shown}
\item{tags}{tags linked to this entity group}
}
}
\section{Fields}{

\describe{
\item{\code{name}}{Name of the \code{\link{VcfFile}}
#' @section Methods:
 \describe{
  \item{\code{has_tag(Tag$new(id=1))}}{Returns true if the VcfFile has is associated with Tag.}
}}
}}

\section{Methods}{

}

\examples{
eg <- VcfFile$new(experiment_id=2)
}
\seealso{
\code{\link{Experiment}}, \code{\link{Entity}}, \code{\link{Specimen}}, \code{\link{VcfFile}} and \code{\link{VcfFile}}

Other data: \code{\link{EntityGroup}},
  \code{\link{Entity}}, \code{\link{Experiment}},
  \code{\link{GenePanel}}, \code{\link{Sample}},
  \code{\link{SnupyGraph}}, \code{\link{SnupyTag}},
  \code{\link{SpecimenProbe}}, \code{\link{VariantDataSet}}
}
\author{
Sebastian Ginzel <sginze2s@inf.h-brs.de>
}
\keyword{datasets}
