% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snupy_data_collection_classes.R
\docType{data}
\name{VcfFileCollection}
\alias{VcfFileCollection}
\title{VcfFilesCollection}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VcfFileCollection
}
\description{
VcfFilesCollection
}
\keyword{datasets}
